pkg/sqlMigrator/file.go
package sqlmigrator

import (
	"bufio"
	"os"
)

func readLastLine(filename string) (string, error) {
	file, err := os.Open(filename)
	if err != nil {
		return "", err
	}
	defer file.Close()

	var lastLine string
	scanner := bufio.NewScanner(file)
	for scanner.Scan() {
		lastLine = scanner.Text()
	}

	if err := scanner.Err(); err != nil {
		return "", err
	}

	return lastLine, nil
}

func createFile(filePath, content string) error {
	file, err := os.Create(filePath)
	if err != nil {
		return err
	}
	defer file.Close()

	if content == "" {
		return nil
	}

	_, err = file.WriteString(content)
	if err != nil {
		return err
	}

	return nil
}
